backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
Django>=5.0,<5.2
django-cors-headers==4.6.0
django-filter==24.3
django-graphql-jwt==0.4.0
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.54.1
frozenlist==1.5.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
kiwisolver==1.4.7
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.2
openai==0.28.0
packaging==24.1
phonenumbers==8.13.51
pillow==11.0.0
promise==2.3
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
wheel
sendgrid==6.11.0